fixup generated file
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@51096 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/router/minidlna/libid3tag-0.15.0b/frametype.h
@@ -37,6 +37,6 @@ extern struct id3_frametype const id3_frametype_unknown;
 extern struct id3_frametype const id3_frametype_obsolete;
 
 struct id3_frametype const *id3_frametype_lookup(register char const *,
-						 register unsigned int);
+						 register size_t);
 
 # endif
